user control

+++ controllers/ControllerAdministration.php
@@ -30,12 +30,7 @@ class ControllerAdministration extends BaseController
         $articles = $this->_articleManager->getAllArticles();
         $this->templateAdmin('views/viewArticleManagement.php',$articles,$title);
     }
-    public function viewArticleManager()
-    {
-        $title = 'Gestion articles';
-        $articles = $this->_articleManager->getAllArticles();
-        echo $this->viewConstruct('views/viewArticleManagement.php',$articles,$title);
-    }
+
     public function userManagement()
     {
         $title = 'Gestion utilisateurs';
@@ -91,9 +86,15 @@ class ControllerAdministration extends BaseController
     {
         $this->_warningManager->deleteWarningByComment($id);
     }
+
     public function deleteUser($id)
     {
-        $this->_userManager->deleteUser($id);
-        echo "utilisateur delete";
+        if($id != $_SESSION['auth']){
+            $this->_userManager->deleteUser($id);
+            echo json_encode($this->_userManager->getAllUsers());
+        }else{
+            echo "impossible";
+        }
+        
     }
 }
